# [GLUTEN-9039][CH] Improve array_sort performance when only single argument is input 