--- a/backends-clickhouse/src/test/scala/org/apache/gluten/execution/GlutenFunctionValidateSuite.scala
+++ b/backends-clickhouse/src/test/scala/org/apache/gluten/execution/GlutenFunctionValidateSuite.scala
@@ -1045,6 +1045,7 @@ class GlutenFunctionValidateSuite extends GlutenClickHouseWholeStageTransformerS
     compareResultsAgainstVanillaSpark(sql, true, { _ => })
   }
 
+<<<<<<< HEAD
   test("GLUTEN-8723 fix slice unexpected exception") {
     val create_sql = "create table t_8723 (full_user_agent string) using orc"
     val insert_sql = "insert into t_8723 values(NULL)"
@@ -1232,7 +1233,8 @@ class GlutenFunctionValidateSuite extends GlutenClickHouseWholeStageTransformerS
   }
 
   test("Test partition values with special characters") {
-    spark.sql("""
+    spark.sql(
+      """
       CREATE TABLE tbl_9050 (
         product_id STRING,
         quantity INT
@@ -1248,5 +1250,15 @@ class GlutenFunctionValidateSuite extends GlutenClickHouseWholeStageTransformerS
     compareResultsAgainstVanillaSpark("select *, input_file_name() from tbl_9050", true, { _ => })
 
     sql("DROP TABLE tbl_9050")
+  }
+
+  test("Test array_sort without comparator") {
+    // default comparator with array elements not nullable guaranteed
+    val sql1 = "select array_sort(split(cast(id * 10 as string), '0')) from range(10)"
+    compareResultsAgainstVanillaSpark(sql1, true, { _ => })
+
+    // default comparator without array elements not nullable guaranteed
+    val sql2 = "select array_sort(array(id+1, null, id+2)) from range(10)"
+    compareResultsAgainstVanillaSpark(sql2, true, { _ => })
   }
 }