## What changes were proposed in this pull request?

(Please fill in changes proposed in this fix)

(Fixes: \#9039)

Before opt: spark  array_sort(arr) ->  CH arraySparkSort(func, arr)
After opt:    spark  array_sort(arr) -> CH  arraySparkSort(arr) 

## How was this patch tested?

(Please explain how this patch was tested. E.g. unit tests, integration tests, manual tests)


(If this patch involves UI changes, please attach a screenshot; otherwise, remove this)