@taiyang-li 2025-03-27
Performance tests: 
```
CREATE TEMPORARY VIEW lineitem
USING org.apache.spark.sql.parquet
OPTIONS (
  path "/data1/liyang/cppproject/gluten/gluten-core/src/test/resources/tpch-data/lineitem"
) ;

select sum(size(x)) from  (select array_sort(split(l_comment, ' ')) as x from lineitem) ;

Before
1 row selected (0.437 seconds)
1 row selected (0.475 seconds)


After
1 row selected (34.186 seconds)
1 row selected (36.021 seconds)

```